third_party/WebKit/LayoutTests/svg/dom/SVGTransformList-expected.txt

This test checks the SVGTransformList API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check passing invalid arguments to 'createSVGTransformFromMatrix'
PASS transform.createSVGTransformFromMatrix() threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': 1 argument required, but only 0 present..
PASS transform.createSVGTransformFromMatrix(svgElement.createSVGTransform()) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS transform.createSVGTransformFromMatrix(svgElement) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS transform.createSVGTransformFromMatrix('aString') threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS transform.createSVGTransformFromMatrix(1) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS transform.createSVGTransformFromMatrix(true) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS transform.createSVGTransformFromMatrix(undefined) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS transform.createSVGTransformFromMatrix(null) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGTransformList': parameter 1 is not of type 'SVGMatrix'..
PASS successfullyParsed is true

TEST COMPLETE
